refactor: move feature-specific utilities and complete shared utils barrel
- Move navigation.ts and handlePositioning.ts to grid-editor/utils
  (only used by grid-editor components and hooks)
- Move split.ts to print-export/utils (only used by print-export)
- Add collision.ts and fill.ts exports to shared/utils barrel for
  consistent import paths
