+++ src/features/grid-editor/utils/handlePositioning.ts
